backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib
bcrypt
python-jose
diffusers
transformers
torch
pillow